requirements.txt
streamlit
pandas
requests
pillow
google-generativeai
openpyxl